+++ vms/client/nx_vms_client_core/unit_tests/src/nx/vms/client/core/resource_views/models/entity_item_model/entity_item_model_test_fixture.cpp
@@ -77,7 +77,7 @@ EntityItemModelTest::RandomNameGenerator EntityItemModelTest::randomNameGenerato
 std::vector<int> EntityItemModelTest::inversePermutation(const std::vector<int>& permutation) const
 {
     std::vector<int> result(permutation.size(), 0);
-    for (int i = 0; i < permutation.size(); ++i)
+    for (int i = 0; i < (int) permutation.size(); ++i)
         result[permutation[i]] = i;
     return result;
 }
